% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_adj_mat}
\alias{calc_adj_mat}
\title{Calculate adjacency matrix from Voronoi cells}
\usage{
calc_adj_mat(pts, cut_vor = TRUE, plot = FALSE, areas = NULL)
}
\arguments{
\item{pts}{Spatial points object}

\item{cut_vor}{Logical, whether to cut Voronoi cells to coastline}

\item{plot}{Logical, whether to plot Voronoi cells}

\item{areas}{Spatial polygons object}
}
\value{
Adjacency matrix
}
\description{
Calculate adjacency matrix from Voronoi cells
}
