% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignotto.R
\name{calc_prop}
\alias{calc_prop}
\title{Calculate proportion of each subset}
\usage{
calc_prop(x)
}
\arguments{
\item{x}{List with counts of points in each subset.}
}
\value{
List with proportions of points in each subset.
}
\description{
Function to calculate proportion of each subset based on the
number of points in each subset.
}
\keyword{internal}
