% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignotto.R
\name{emp_kl_div}
\alias{emp_kl_div}
\title{Empirical KL divergence between two locations}
\usage{
emp_kl_div(x, y, prob = 0.9)
}
\arguments{
\item{x}{Vector of data representing location 1}

\item{y}{Vector of data representing location 2}

\item{prob}{Extremal quantile to use.}
}
\value{
KL divergence between x and y.
}
\description{
Functions to calculate empirical KL divergence between two locations using
the Vignotto 2021 method.
}
\keyword{internal}
