% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{fit_ce}
\alias{fit_ce}
\title{Fit Conditional Extremes model}
\usage{
fit_ce(
  data,
  vars = c("rain", "wind_speed"),
  marg_prob = 0.9,
  cond_prob = 0.9,
  f = list(excess ~ name, ~1),
  split_data = TRUE
)
}
\arguments{
\item{data}{List of dataframes at each location containing data for each
variable.}

\item{vars}{Variable names for each location.}

\item{marg_prob}{Marginal quantile for thresholding.}

\item{cond_prob}{Conditional quantile for dependence modelling.}

\item{f}{Formula for \code{evgam} model.}

\item{split_data}{if \code{data} has variables stacked, unstack.}
}
\value{
Object of type \code{mexDependence} for each location.
}
\description{
Fit Conditional Extremes model using marginal fits form \code{evgam}.
}
