% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{fit_texmex_dep}
\alias{fit_texmex_dep}
\title{Fit \code{texmex} dependence model}
\usage{
fit_texmex_dep(
  marginal,
  vars = c("rain", "wind_speed"),
  mex_dep_args = list(start = c(0.01, 0.01), dqu = 0.7, fixed_b = FALSE, PlotLikDo =
    FALSE),
  fit_no_keef = FALSE
)
}
\arguments{
\item{marginal}{List of \code{migpd} objects for each site.}

\item{vars}{Variables to fit dependence on.}

\item{mex_dep_args}{Arguments to pass to \link[texmex]{mexDependence}.}

\item{fit_no_keef}{If model doesn't fit under Keef constraints, fit without
(see \link[texmex]{mexDependence} for details).}
}
\value{
List of \code{mexDependence} objects for each site.
}
\description{
Fit \code{texmex} dependence model for each site.
}
\keyword{internal}
