% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{gen_marg_migpd}
\alias{gen_marg_migpd}
\title{Generate marginal \code{migpd} objects}
\usage{
gen_marg_migpd(data_gpd, data, mqu = 0.95)
}
\arguments{
\item{data_gpd}{Scale and shape parameters for each location.}

\item{data}{Data for each location.}

\item{mqu}{Marginal quantile for thresholding.}
}
\value{
List of \code{migpd} objects for each location.
}
\description{
Generate marginal \code{migpd} objects from \code{evgam} objects for each
site.
}
