% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_clust.R
\name{js_div}
\alias{js_div}
\title{Calculate Jensen-Shannon divergence for each data point}
\usage{
js_div(params_x, params_y, thresh_max, data_max = 2 * thresh_max, n_dat = 10)
}
\arguments{
\item{params_x}{Parameters for the first Gaussian distribution.}

\item{params_y}{Parameters for the second Gaussian distribution.}

\item{thresh_max}{Maximum threshold value across all locations.}

\item{data_max}{Maximum data value.}

\item{n_dat}{Number of data points to use in the Jensen-Shannon divergence.
return The Jensen-Shannon divergence for each data point.}
}
\description{
Function to calculate Jensen-Shannon divergence for each data
point.
}
\keyword{internal}
