% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_clust.R
\name{js_gauss}
\alias{js_gauss}
\title{Calculate Jensen-Shannon divergence between two Gaussian distributions}
\usage{
js_gauss(mu1, mu2, var1, var2)
}
\arguments{
\item{mu1}{Mean of the first Gaussian distribution.}

\item{mu2}{Mean of the second Gaussian distribution.}

\item{var1}{Variance of the first Gaussian distribution.}

\item{var2}{Variance of the second Gaussian distribution.}
}
\value{
The Jensen-Shannon divergence between the two Gaussian
distributions.
}
\description{
Function to calculate the Jensen-Shannon divergence between two
Gaussian distributions, using the Kullback-Leibler divergence between them.
This metric is symmetric, as required for clustering.
}
\keyword{internal}
