% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_clust.R
\name{kl_gauss}
\alias{kl_gauss}
\title{Calculate KL divergence between two Gaussian distributions}
\usage{
kl_gauss(mu1, mu2, var1, var2)
}
\arguments{
\item{mu1}{Mean of the first Gaussian distribution.}

\item{mu2}{Mean of the second Gaussian distribution.}

\item{var1}{Variance of the first Gaussian distribution.}

\item{var2}{Variance of the second Gaussian distribution.}
}
\value{
The Kullback-Leibler divergence between the two Gaussian
distributions.
}
\description{
Function to calculate the Kullback-Leibler divergence between
two Gaussian distributions, using the closed form solution for the case of
two univariate Gaussians provided at
\href{https://statproofbook.github.io/P/norm-kl.html}{test}.
}
\keyword{internal}
