% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignotto.R
\name{kl_sim_eval}
\alias{kl_sim_eval}
\title{PAM clustering based on empirical KL divergence}
\usage{
kl_sim_eval(data, kl_prob, k = NULL, kl_mat = NULL, cluster_mem = NULL, ...)
}
\arguments{
\item{data}{Data matrix with rows representing locations and columns}

\item{kl_prob}{Extremal quantile to use.}

\item{k}{Number of clusters, set to NULL to produce scree plot.}

\item{kl_mat}{Distance matrix, set to NULL to calculate.}

\item{cluster_mem}{Optional true cluster membership to calculate adjusted
Rand index.}

\item{...}{Additional arguments to pass to \code{proxy::dist}.}
}
\value{
List with clustering solution and adjusted Rand index if cluster_mem
is provided.
}
\description{
Perform clustering based on empirical KL divergence between
bivariate extremes for two locations.
}
