\encoding{UTF-8}

\name{Medians for ecological indicator values}
\alias{moeller}

\title{
Calculate Median of Ecological Indicator Values.
}

\description{
Calculate the median of ecological indicator values according to the method of Möller (1992)
}

\usage{
moeller(x)
}

\arguments{
  \item{ x }{
    numeric. In the typical case integer values.
}
}

\details{
Nothing here yet.
}

\value{
\code{numeric} vector.
}

\references{
Möller, H (2013). Zur Verwendung des Medians bei Zeigerwertberechnungen nach ELLENBERG. Tuexenia, 12:25-28
}

\author{
Roland Kaiser
}

\seealso{
\code{\link{eiv}}
}

\keyword{ database }

