% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignotto.R
\name{pareto_trans}
\alias{pareto_trans}
\title{Pareto transformation}
\usage{
pareto_trans(x)
}
\arguments{
\item{x}{Vector to convert.}
}
\value{
Vector in Pareto scale.
}
\description{
Function to convert a vector to Pareto scale.
}
\keyword{internal}
