% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignotto.R
\name{partition_max}
\alias{partition_max}
\title{Partition into 3 subsets}
\usage{
partition_max(x, prob, plot = FALSE)
}
\arguments{
\item{x}{List of vectors representing different locations.}

\item{prob}{Extremal quantile to use.}
}
\value{
List with counts of points in each subset.
}
\description{
Function to partition bivariate extremes into into 3 subsets
based on the risk function from Vignotto 2021.
}
\keyword{internal}
