% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plt_clust_map}
\alias{plt_clust_map}
\title{Plot clustering solution on map}
\usage{
plt_clust_map(pts, areas, clust_obj)
}
\arguments{
\item{pts}{Spatial points object}

\item{areas}{Spatial polygons object}

\item{clust_obj}{Clustering object}
}
\value{
ggplot object
}
\description{
Plot clustering solution on map
}
