% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plt_sil_map}
\alias{plt_sil_map}
\title{Plot silhouette width on map}
\usage{
plt_sil_map(pts, areas, sil_obj, medoids = NULL)
}
\arguments{
\item{pts}{Spatial points object}

\item{areas}{Spatial polygons object}

\item{sil_obj}{silhouette object}

\item{medoids}{Medoid names, Default: NULL.}
}
\value{
ggplot object
}
\description{
Plot silhouette width for clustering solution on map.
}
