% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_clust.R
\name{pull_thresh_trans}
\alias{pull_thresh_trans}
\title{Pull thresholds for conditional extremes model}
\usage{
pull_thresh_trans(dep)
}
\arguments{
\item{dep}{List of \code{mexDependence} objects for each location.}
}
\value{
List of named vectors containing the thresholds for each location.
}
\description{
Function to pull the thresholds for the conditional extremes
model.
}
\keyword{internal}
