% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignotto.R
\name{risk_fun}
\alias{risk_fun}
\title{Bivariate risk function}
\usage{
risk_fun(x, fun = max)
}
\arguments{
\item{x}{List of vectors representing different locations.}

\item{fun}{Function to calculate "risk", defaults to max as easier to
partition. Note that \code{sum} has not been implemented yet.}
}
\value{
Vector of risk values.
}
\description{
Function to calculate bivariate risk function from Vignotto
2021.
}
\keyword{internal}
