% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree plot}
\usage{
scree_plot(dist_mat, k = 1:10, fun = cluster::pam, ...)
}
\arguments{
\item{dist_mat}{Distance matrix}

\item{k}{Number of clusters}

\item{fun}{Clustering function}

\item{...}{Additional arguments to clustering function}
}
\value{
Total within-cluster sum of distances
}
\description{
Produce scree plot
}
