% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sil_boxplot}
\alias{sil_boxplot}
\title{Silhouette boxplot}
\usage{
sil_boxplot(
  dist_mat,
  k = 2:10,
  fun = cluster::pam,
  show_plt = TRUE,
  ret_sil = TRUE,
  ...
)
}
\arguments{
\item{dist_mat}{Distance matrix}

\item{k}{Number of clusters}

\item{fun}{Clustering function, Default: \code{cluster::pam}}

\item{show_plt}{Logical, whether to show plot}

\item{ret_sil}{Logical, whether to return dataframe of silhouette values}

\item{...}{Additional arguments to clustering function}
}
\value{
ggplot object
}
\description{
Produce boxplot of silhouette widths for different values of k.
}
