% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{within_cluster_sum}
\alias{within_cluster_sum}
\title{Calculate within cluster sum of distances}
\usage{
within_cluster_sum(k, distance_matrix, fun = cluster::pam, ...)
}
\arguments{
\item{k}{Number of clusters}

\item{distance_matrix}{Distance matrix}

\item{fun}{Clustering function}
}
\value{
Total within-cluster sum of distances
}
\description{
Calculate within cluster sum of distances
}
\keyword{internal}
