% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_clust.R
\name{js_clust}
\alias{js_clust}
\title{Cluster based on Jensen-Shannon divergence for conditional extremes
model}
\usage{
js_clust(
  dependence,
  nclust = 3,
  cluster_mem = NULL,
  dat_max_mult = 2,
  n_dat = 10
)
}
\arguments{
\item{dependence}{List of \code{mexDependence} objects for each location.}

\item{nclust}{Number of clusters to fit.}

\item{cluster_mem}{Optional known cluster memberships for each location,
which can be used to evaluate the quality of the clustering.}

\item{dat_max_mult}{Multiplier for the maximum multiple of the largest
threshold across all locations.}

\item{n_dat}{Number of data points to use in the Jensen-Shannon divergence.}
}
\value{
List containing the clustering results and, if \code{cluster_mem} is
provided, the adjusted Rand index.
}
\description{
Function to calculate and cluster on the Jensen-Shannon
Divergence for the conditional extremes model.
}
